# src/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.dev"
  title: "Dart REST API"
  description: |
    *DRAFT*

    Dart, short for "data mart à la carte", is a self-service data workflow solution that
    facilitates the creation, transformation, and movement of data using data engines,
    primarily in the context of AWS. Put another way, it is a system to facilitate moving
    data in and out of datastores like S3, RDS, Redshift, EMR, DynamoDB, something custom, etc.

    This document describes the Dart REST API using the Swagger 2.0 format. Portions of
    the API are not fully documented due to limitations of the Swagger 2.0 specification
    and the complexity of this API. These APIs will be marked with *NOTE* and will provide
    more details in the descriptive text.
  contact:
    url: https://github.com/RetailMeNotSandbox/dart
  license:
    name: MIT
    url: https://raw.githubusercontent.com/RetailMeNotSandbox/dart/master/LICENSE.txt
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
  /action:
    get:
      summary: "returns the actions matching the filters or all actions if no filters are specified"
      tags:
        - "action"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
        - $ref: '#/parameters/orderByParam'
        - name: "datastore_id"
          type: "string"
          in: "query"
        - name: "workflow_id"
          type: "string"
          in: "query"
      responses:
        "200":
          description: "returns the actions matching the filters or all actions if no filters are specified."
          schema:
            type: object
            properties:
              results:
                type: "array"
                items:
                  $ref: "#/definitions/Action"
              limit:
                type: "integer"
              offset:
                type: "integer"
              total:
                type: "integer"
  /action/{action_id}:
    parameters:
      - name: "action_id"
        type: "string"
        in: "path"
        required: true
    get:
      summary: "return the action"
      tags:
        - "action"
      responses:
        "200":
          description: returns the action
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Action"
    put:
      summary: "update the action"
      tags:
        - "action"
      parameters:
        - name: "action"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Action"
      responses:
        "200":
          description: returns the action
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Action"
    patch:
      summary: "patch the action using JSON Patch"
      tags:
        - "action"
      parameters:
        - $ref: '#/parameters/patchParam'
      responses:
        "200":
          description: returns the action
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Action"
    delete:
      summary: "delete the action"
      tags:
        - "action"
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResult"
...
  /dataset:
    get:
      summary: "returns the datasets matching the filters or all datasets if no filters are specified"
      tags:
        - "dataset"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
      responses:
        "200":
          description: "returns the datasets matching the filters or all datasets if no filters are specified."
          schema:
            type: object
            properties:
              results:
                type: "array"
                items:
                  $ref: "#/definitions/Dataset"
              limit:
                type: "integer"
              offset:
                type: "integer"
              total:
                type: "integer"
    post:
      summary: "create a dataset"
      tags:
        - "dataset"
      parameters:
        - name: "dataset"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Dataset"
      responses:
        "200":
          description: returns the created Dataset
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Dataset"
  /dataset/guess:
    get:
      summary: "best guess of the dataset"
      tags:
        - "dataset"
      parameters:
        - name: s3_path
          in: query
          type: string
          required: true
        - name: max_lines
          in: query
          type: integer
          required: true
      responses:
        "200":
          description: returns the dataset
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Dataset"
  /dataset/{dataset_id}:
    parameters:
      - name: "dataset_id"
        type: "string"
        in: "path"
        required: true
    get:
      summary: "get a dataset"
      tags:
        - "dataset"
      responses:
        "200":
          description: returns the dataset
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Dataset"
    put:
      summary: "update a dataset"
      tags:
        - "dataset"
      parameters:
        - name: "dataset"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Dataset"
      responses:
        "200":
          description: returns the dataset
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Dataset"
    delete:
      summary: "delete a dataset"
      tags:
        - "dataset"
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResult"
...
  /datastore:
    post:
      summary: "create a datastore"
      tags:
        - "datastore"
      parameters:
        - name: "datastore"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Datastore"
      responses:
        "200":
          description: returns the created datastore
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Datastore"
    get:
      summary: "find datastores"
      tags:
        - "datastore"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
      responses:
        "200":
          description: "returns the datastores matching the filters or all datastores if no filters are specified."
          schema:
            type: object
            properties:
              results:
                type: "array"
                items:
                  $ref: "#/definitions/Datastore"
              limit:
                type: "integer"
              offset:
                type: "integer"
              total:
                type: "integer"
  /datastore/{datastore_id}:
    parameters:
      - name: "datastore_id"
        type: "string"
        in: "path"
        required: true
    get:
      summary: "return the datastore"
      tags:
        - "datastore"
      responses:
        "200":
          description: returns the datastore
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Datastore"
    put:
      summary: "update the datastore"
      tags:
        - "datastore"
      parameters:
        - name: "datastore"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Datastore"
      responses:
        "200":
          description: returns the datastore
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Datastore"
        "400":
          description: "datastore is not in a valid state"
          schema:
            $ref: "#/definitions/ErrorResult"
    patch:
      summary: "patch the datastore using JSON Patch"
      tags:
        - "datastore"
      parameters:
        - $ref: '#/parameters/patchParam'
      responses:
        "200":
          description: returns the datastore
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Datastore"
        "400":
          description: "datastore is not in a valid state"
          schema:
            $ref: "#/definitions/ErrorResult"
    delete:
      summary: "delete the datastore"
      tags:
        - "datastore"
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResult"
  /datastore/{datastore_id}/action:
    parameters:
      - name: "datastore_id"
        type: "string"
        in: "path"
        required: true
    post:
      summary: "create an action for a datastore"
      tags:
        - "datastore"
        - "action"
      parameters:
        - name: "actions"
          in: "body"
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/Action"
          description: |
            ~~~~
            NOTE: This parameter is specified as an array because Swagger 2.0 does
            not support the oneOf construct. This API also supports passing a single
            Action object.
            ~~~~
      responses:
        "200":
          description: "the created actions"
          schema:
            type: object
            properties:
              results:
                type: array
                items:
                  $ref: "#/definitions/Action"
...
  /workflow/{workflow_id}/action:
    parameters:
      - name: "workflow_id"
        description: "the workflow id"
        type: "string"
        in: "path"
        required: true
    post:
      summary: "create an action for a workflow"
      tags:
        - "workflow"
        - "action"
      parameters:
        - name: "actions"
          in: "body"
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/Action"
          description: |
            ~~~~
            NOTE: This parameter is specified as an array because Swagger 2.0 does
            not support the oneOf construct. This API also supports passing a single
            Action object.
            ~~~~
      responses:
        "200":
          description: "the created actions"
          schema:
            type: object
            properties:
              results:
                type: array
                items:
                  $ref: "#/definitions/Action"
...
parameters:
  limitParam:
    name: "limit"
    type: "integer"
    in: "query"
    description: |
      Specify the maximum number of records that should be returned by this request.
  offsetParam:
    name: "offset"
    type: "integer"
    in: "query"
    description: |
      Specify the starting index for the next chunk of results to be returned by this request.
  filtersParam:
    name: filters
    in: query
    type: string
    description: |
      ~~~~
      NOTE: This parameter is specified as a string because Swagger 2.0 does
      not support passing object or array types as query string parameters.
      ~~~~
      To use, pass a URL encoded JSON array of Filter objects (as defined in
      the definitions section).
  orderByParam:
    name: "order_by"
    type: string
    in: query
    description: |
      ~~~~
      NOTE: This parameter is specified as a string because Swagger 2.0 does
      not support passing object or array types as query string parameters.
      ~~~~
      To use, pass a URL encoded JSON array of OrderBy objects (as defined in
      the definitions section).
  patchParam:
    name: "patch"
    description: "a JSON Patch object"
    in: "body"
    required: true
    schema:
      $ref: '#/definitions/JSONPatch'
#
# Definitions are auto-generated by dart.tool.export_swagger_definitions
# Do not manually edit!
#
definitions:
  Action:
    additionalProperties: false
    properties:
      created:
        type: string
      data:
        $ref: '#/definitions/ActionData'
      id:
        type: string
      updated:
        type: string
      version_id:
        type: integer
    required:
    - data
    type: object
...
  ActionData:
    additionalProperties: false
    properties:
      action_type_name:
        type: string
      args:
        additionalProperties: false
        properties: {}
        type: object
      datastore_id:
        type: string
      ecs_task_arn:
        type: string
      end_time:
        type: string
      engine_name:
        type: string
      error_message:
        type: string
        x-schema-form:
          type: textarea
      extra_data:
        type: object
      first_in_workflow:
        default: false
        type: boolean
      last_in_workflow:
        default: false
        type: boolean
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      on_failure:
        default: DEACTIVATE
        description: applies to the workflow if this is a workflow action template,
          otherwise the datastore
        enum:
        - DEACTIVATE
        - CONTINUE
        type: string
      on_failure_email:
        default: []
        items:
          pattern: ^\S+@\S+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''email '' + $index }}'
          type: tabarray
      on_success_email:
        default: []
        items:
          pattern: ^\S+@\S+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''email '' + $index }}'
          type: tabarray
      order_idx:
        minimum: 0.0
        type: number
      progress:
        type: number
      queued_time:
        type: string
      start_time:
        type: string
      state:
        default: HAS_NEVER_RUN
        enum:
        - HAS_NEVER_RUN
        - QUEUED
        - PENDING
        - RUNNING
        - FINISHING
        - COMPLETED
        - FAILED
        - TEMPLATE
        type: string
      tags:
        default: []
        items:
          maxLength: 30
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
      workflow_action_id:
        type: string
      workflow_id:
        type: string
      workflow_instance_id:
        type: string
    required:
    - action_type_name
    - engine_name
    - name
    type: object
...
  Column:
    additionalProperties: false
    properties:
      data_type:
        enum:
        - STRING
        - VARCHAR
        - BIGINT
        - INT
        - SMALLINT
        - DOUBLE
        - FLOAT
        - NUMERIC
        - BOOLEAN
        - DATE
        - DATETIME
        - TIMESTAMP
        type: string
      date_pattern:
        type: string
      description:
        type: string
      is_nullable:
        default: true
        type: boolean
      length:
        minimum: 1
        type: integer
      name:
        maxLength: 255
        pattern: ^[a-zA-Z0-9_]+$
        type: string
      path:
        type: string
      precision:
        type: integer
      scale:
        type: integer
    required:
    - name
    - data_type
    type: object
  DataFormat:
    additionalProperties: false
    properties:
      delimited_by:
        type: string
      escaped_by:
        type: string
      file_format:
        enum:
        - TEXTFILE
        - PARQUET
        - RCFILE
        type: string
      null_string:
        type: string
      num_header_rows:
        default: 0
        minimum: 0
        type: integer
      quoted_by:
        type: string
      regex_input:
        type: string
      regex_output:
        type: string
      row_format:
        enum:
        - DELIMITED
        - JSON
        - REGEX
        - NONE
        type: string
    required:
    - file_format
    - row_format
    type: object
  Dataset:
    additionalProperties: false
    properties:
      created:
        type: string
      data:
        $ref: '#/definitions/DatasetData'
      id:
        type: string
      updated:
        type: string
      version_id:
        type: integer
    required:
    - data
    type: object
  DatasetData:
    additionalProperties: false
    properties:
      batch_merge_sort_keys:
        default: []
        items:
          maxLength: 127
          pattern: ^[a-zA-Z0-9_]+ (ASC|DESC)$
          type: string
        type: array
        x-schema-form:
          title: '{{ value || ''upsert_sort_key '' + $index }}'
          type: tabarray
      columns:
        items:
          $ref: '#/definitions/Column'
        minItems: 1
        type: array
        x-schema-form:
          title: '{{ value.name || ''column '' + $index }}'
          type: tabarray
      compression:
        enum:
        - NONE
        - BZ2
        - GZIP
        - SNAPPY
        type: string
      data_format:
        $ref: '#/definitions/DataFormat'
      description:
        type: string
      distribution_keys:
        default: []
        items:
          maxLength: 127
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        type: array
        x-schema-form:
          title: '{{ value || ''distribution_key '' + $index }}'
          type: tabarray
      hive_compatible_partition_folders:
        default: false
        type: boolean
      load_type:
        enum:
        - INSERT
        - MERGE
        - RELOAD_ALL
        - RELOAD_LAST
        type: string
      location:
        pattern: ^s3://.+$
        type: string
      merge_keys:
        default: []
        items:
          maxLength: 127
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        type: array
        x-schema-form:
          title: '{{ value || ''merge_key '' + $index }}'
          type: tabarray
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      partitions:
        items:
          $ref: '#/definitions/Column'
        minItems: 0
        type:
        - array
        - 'null'
        x-schema-form:
          title: '{{ value.name || ''column '' + $index }}'
          type: tabarray
      primary_keys:
        default: []
        items:
          maxLength: 127
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        type: array
        x-schema-form:
          title: '{{ value || ''primary_key '' + $index }}'
          type: tabarray
      sort_keys:
        default: []
        items:
          maxLength: 127
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        type: array
        x-schema-form:
          title: '{{ value || ''sort_key '' + $index }}'
          type: tabarray
      table_name:
        minLength: 1
        pattern: ^[a-zA-Z0-9_]+$
        type: string
      tags:
        default: []
        items:
          maxLength: 30
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
    required:
    - name
    - table_name
    - location
    - load_type
    - data_format
    - columns
    - compression
    type: object
  Datastore:
    additionalProperties: false
    properties:
      created:
        type: string
      data:
        $ref: '#/definitions/DatastoreData'
      id:
        type: string
      updated:
        type: string
      version_id:
        type: integer
    required:
    - data
    type: object
  DatastoreData:
    additionalProperties: false
    properties:
      args:
        type: 'null'
      concurrency:
        default: 1
        maximum: 10
        minimum: 1
        type: integer
      connection_url:
        type: string
      engine_name:
        pattern: ^[a-zA-Z0-9_]+$
        type: string
      extra_data:
        type: object
      host:
        type: string
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      port:
        type: integer
      s3_artifacts_path:
        type: string
      s3_logs_path:
        type: string
      state:
        default: INACTIVE
        enum:
        - INACTIVE
        - ACTIVE
        - DONE
        - TEMPLATE
        type: string
      tags:
        default: []
        items:
          maxLength: 30
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
      workflow_datastore_id:
        type: string
    required:
    - name
    - engine_name
    type: object
...
  ErrorResult:
    properties:
      error_message:
        type: string
      results:
        enum:
        - ERROR
        type: string
    type: object
...
  JSONPatch:
    description: JSON Patch object as defined by http://json.schemastore.org/json-patch
    type: object
...
  OKResult:
    properties:
      results:
        enum:
        - OK
        type: string
    type: object
...